# Conflicts:
#	src/main/java/com/project/uandmeet/controller/BoardCotroller.java
#	src/main/java/com/project/uandmeet/repository/EntryRepository.java
#	src/main/java/com/project/uandmeet/repository/MemberRepostiory.java
#	src/main/java/com/project/uandmeet/service/BoardService.java

+++ src/main/java/com/project/uandmeet/service/BoardService.java
@@ -5,13 +5,12 @@ import com.project.uandmeet.Exception.ErrorCode;
 import com.project.uandmeet.dto.MemberDtoGroup.MemberSimpleDto;
 import com.project.uandmeet.dto.boardDtoGroup.BoardRequestDto;
 import com.project.uandmeet.dto.boardDtoGroup.BoardResponseDto;
+import com.project.uandmeet.dto.boardDtoGroup.LikeDto;
 import com.project.uandmeet.model.Board;
 import com.project.uandmeet.model.Category;
+import com.project.uandmeet.model.Like;
 import com.project.uandmeet.model.Member;
-import com.project.uandmeet.repository.BoardRepository;
-import com.project.uandmeet.repository.CategoryRepository;
-import com.project.uandmeet.repository.EntryRepository;
-import com.project.uandmeet.repository.MemberRepostiory;
+import com.project.uandmeet.repository.*;
 import com.project.uandmeet.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,11 +27,11 @@ public class BoardService {
     private final MemberRepostiory memberRepostiory;
     private final CategoryRepository categoryRepository;
 
-    private final EntryRepository entryRepository;
+    private final LikeRepository likeRepository;
 
     //게시판 생성
     @Transactional
-    public CustomException boardNew(BoardRequestDto boardRequestDto, UserDetailsImpl userDetails) {
+    public CustomException boardNew(BoardRequestDto.createAndCheck boardRequestDto, UserDetailsImpl userDetails) {
         //로그인 유저 정보.
         Member memberTemp = memberRepostiory.findById(userDetails.getMember().getId())
                 .orElseThrow(()->new CustomException(ErrorCode.EMPTY_CONTENT));
@@ -41,9 +40,14 @@ public class BoardService {
               .orElseThrow(()->new CustomException(ErrorCode.EMPTY_CONTENT));
 
         Board board = new Board(memberTemp,category, boardRequestDto);
-        boardRepository.save(board);
 
-        return new CustomException(ErrorCode.COMPLETED_OK);
+        try {
+            boardRepository.save(board);
+            return new CustomException(ErrorCode.COMPLETED_OK);
+        }catch (Exception e){
+            System.out.println(e);
+            return new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
+        }
     }
 
     //게시물 전체 조회
@@ -59,24 +63,22 @@ public class BoardService {
             for (Board boardTemp : boards) {
                 //작성자 간이 닉네임 생성.
                 MemberSimpleDto memberSimpleDto = new MemberSimpleDto(boardTemp.getMember().getNickname(),
-                        boardTemp.getMember().getUsername(), boardTemp.getMember().getProfile());
+                        boardTemp.getMember().getEmail(), boardTemp.getMember().getProfile());
 
                 BoardResponseDto boardResponseDto = new BoardResponseDto(memberSimpleDto,boardTemp);
                 boardResponseDtos.add(boardResponseDto);
             }
         }
-
         return boardResponseDtos;
     }
 
-    @Transactional
     //게시물 상세 조회
+    @Transactional
     public BoardResponseDto boardChoiceInquiry(Long id) {
 
         //개시판 정보 추출
         Board boards = boardRepository.findById(id)
                 .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
-
 
         // 찾으 정보를 Dto로 변환 한다.
         BoardResponseDto boardResponseDto = null;
@@ -84,7 +86,7 @@ public class BoardService {
         if (boards != null) {
             //작성자 간이 닉네임 생성.
             MemberSimpleDto memberSimpleDto = new MemberSimpleDto(boards.getMember().getNickname(),
-                    boards.getMember().getUsername(), boards.getMember().getProfile());
+                    boards.getMember().getEmail(), boards.getMember().getProfile());
 
             boardResponseDto = new BoardResponseDto(memberSimpleDto, boards);
             return boardResponseDto;
@@ -92,17 +94,69 @@ public class BoardService {
         else return null;
     }
 
+    //게시물 삭제.
     @Transactional
     public CustomException boardDel(Long id, UserDetailsImpl userDetails) {
+
         //로그인 유저 정보.
-        Member MemberTemp = memberRepostiory.findById(userDetails.getMember().getId())
-                .orElseThrow(()->new CustomException(ErrorCode.EMPTY_CONTENT));
+        Member memberTemp = memberRepostiory.findById(userDetails.getMember().getId())
+                .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
+        Board board = boardRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
 
-        boardRepository.deleteById(id);
-
-        return new CustomException(ErrorCode.COMPLETED_OK);
+        //본인이 아니면 예외처리
+        if (board.getMember().getEmail().equals(memberTemp.getEmail())) {
+            try {
+                boardRepository.deleteById(id);
+                return new CustomException(ErrorCode.COMPLETED_OK);
+            } catch (Exception e) {
+                System.out.println(e);
+                return new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
+            }
+        }
+        else {
+            return new CustomException(ErrorCode.INVALID_AUTHORITY);
+        }
     }
 
-    //게시물 삭제.
+    //게시물 수정
+    @Transactional
+    public CustomException boardUpdate(Long id, BoardRequestDto.createAndCheck boardRequestUdateDto,
+                                       UserDetailsImpl userDetails) {
 
+        //로그인 유저 정보.
+        Member memberTemp = memberRepostiory.findById(userDetails.getMember().getId())
+                .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
+
+        Board board = boardRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
+
+        //본인이 아니면 예외처리
+        if (board.getMember().getEmail().equals(memberTemp.getEmail())) {
+            Board boardUpdate = new Board(board, boardRequestUdateDto);
+            try {
+                boardRepository.save(boardUpdate);
+                return new CustomException(ErrorCode.COMPLETED_OK);
+            } catch (Exception e) {
+                System.out.println(e);
+                return new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
+            }
+        } else {
+            return new CustomException(ErrorCode.INVALID_AUTHORITY);
+        }
+    }
+
+
+    //수정중
+    @Transactional
+    public CustomException likeClick(LikeDto likeDto, UserDetailsImpl userDetails) {
+
+        Member memberTemp = memberRepostiory.findById(userDetails.getMember().getId())
+                .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
+
+        Board board = boardRepository.findById(likeDto.getPostid())
+                .orElseThrow(() -> new CustomException(ErrorCode.EMPTY_CONTENT));
+
+        return null;
+    }
 }
